IGNITE-15457 Moves thread pool metrics registration to PoolProcessor.
