+++ modules/core/src/main/java/org/apache/ignite/internal/IgniteKernal.java
@@ -1100,12 +1100,10 @@ public class IgniteKernal implements IgniteEx, IgniteMXBean, Externalizable {
             // Start security processors.
             startProcessor(securityProcessor());
 
-            PoolProcessor pools = new PoolProcessor(ctx);
-
-            startProcessor(pools);
+            startProcessor(new PoolProcessor(ctx));
 
             // Run background network diagnostics.
-            GridDiagnostic.runBackgroundCheck(igniteInstanceName, pools.getExecutorService(), log);
+            GridDiagnostic.runBackgroundCheck(igniteInstanceName, ctx.pools().getExecutorService(), log);
 
             // Closure processor should be started before all others
             // (except for resource processor), as many components can depend on it.
@@ -1325,53 +1323,12 @@ public class IgniteKernal implements IgniteEx, IgniteMXBean, Externalizable {
 
             startTimer.finishGlobalStage("Await transition");
 
-            ctx.metric().registerThreadPools(
-                pools.utilityCachePool(),
-                pools.getExecutorService(),
-                pools.getServiceExecutorService(),
-                pools.getSystemExecutorService(),
-                pools.getStripedExecutorService(),
-                pools.getPeerClassLoadingExecutorService(),
-                pools.getManagementExecutorService(),
-                pools.getDataStreamerExecutorService(),
-                pools.getRestExecutorService(),
-                pools.getAffinityExecutorService(),
-                pools.getIndexingExecutorService(),
-                pools.asyncCallbackPool(),
-                pools.getQueryExecutorService(),
-                pools.getSchemaExecutorService(),
-                pools.getRebalanceExecutorService(),
-                pools.getStripedRebalanceExecutorService(),
-                pools.customExecutors());
-
             registerMetrics();
 
             ctx.cluster().registerMetrics();
 
             // Register MBeans.
-            mBeansMgr.registerMBeansAfterNodeStarted(
-                pools.utilityCachePool(),
-                pools.getExecutorService(),
-                pools.getServiceExecutorService(),
-                pools.getSystemExecutorService(),
-                pools.getStripedExecutorService(),
-                pools.getPeerClassLoadingExecutorService(),
-                pools.getManagementExecutorService(),
-                pools.getDataStreamerExecutorService(),
-                pools.getRestExecutorService(),
-                pools.getAffinityExecutorService(),
-                pools.getIndexingExecutorService(),
-                pools.asyncCallbackPool(),
-                pools.getQueryExecutorService(),
-                pools.getSchemaExecutorService(),
-                pools.getRebalanceExecutorService(),
-                pools.getStripedRebalanceExecutorService(),
-                pools.customExecutors(),
-                ctx.workersRegistry());
-
-            ctx.systemView().registerThreadPools(
-                pools.getStripedExecutorService(),
-                pools.getDataStreamerExecutorService());
+            mBeansMgr.registerMBeansAfterNodeStarted();
 
             boolean recon = false;
