+++ modules/core/src/main/java/org/apache/ignite/internal/managers/IgniteMBeansManager.java
@@ -18,9 +18,7 @@
 package org.apache.ignite.internal.managers;
 
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ExecutorService;
 import javax.management.JMException;
 import javax.management.ObjectName;
 import org.apache.ignite.IgniteCheckedException;
@@ -32,8 +30,6 @@ import org.apache.ignite.internal.GridKernalContextImpl;
 import org.apache.ignite.internal.IgniteKernal;
 import org.apache.ignite.internal.QueryMXBeanImpl;
 import org.apache.ignite.internal.ServiceMXBeanImpl;
-import org.apache.ignite.internal.StripedExecutorMXBeanAdapter;
-import org.apache.ignite.internal.ThreadPoolMXBeanAdapter;
 import org.apache.ignite.internal.TransactionMetricsMxBeanImpl;
 import org.apache.ignite.internal.TransactionsMXBeanImpl;
 import org.apache.ignite.internal.managers.encryption.EncryptionMXBeanImpl;
@@ -44,11 +40,9 @@ import org.apache.ignite.internal.processors.cache.warmup.WarmUpMXBeanImpl;
 import org.apache.ignite.internal.processors.cluster.BaselineAutoAdjustMXBeanImpl;
 import org.apache.ignite.internal.processors.metric.MetricsMxBeanImpl;
 import org.apache.ignite.internal.processors.performancestatistics.PerformanceStatisticsMBeanImpl;
-import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.internal.worker.FailureHandlingMxBeanImpl;
 import org.apache.ignite.internal.worker.WorkersControlMXBeanImpl;
-import org.apache.ignite.internal.worker.WorkersRegistry;
 import org.apache.ignite.mxbean.BaselineAutoAdjustMXBean;
 import org.apache.ignite.mxbean.ClusterMetricsMXBean;
 import org.apache.ignite.mxbean.ComputeMXBean;
@@ -62,14 +56,10 @@ import org.apache.ignite.mxbean.PerformanceStatisticsMBean;
 import org.apache.ignite.mxbean.QueryMXBean;
 import org.apache.ignite.mxbean.ServiceMXBean;
 import org.apache.ignite.mxbean.SnapshotMXBean;
-import org.apache.ignite.mxbean.StripedExecutorMXBean;
-import org.apache.ignite.mxbean.ThreadPoolMXBean;
 import org.apache.ignite.mxbean.TransactionMetricsMxBean;
 import org.apache.ignite.mxbean.TransactionsMXBean;
 import org.apache.ignite.mxbean.WarmUpMXBean;
 import org.apache.ignite.mxbean.WorkersControlMXBean;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
-import org.jetbrains.annotations.Nullable;
 
 /**
  * Class that registers and unregisters MBeans for kernal.
@@ -99,46 +89,9 @@ public class IgniteMBeansManager {
     /**
      * Registers kernal MBeans (for kernal, metrics, thread pools) after node start.
      *
-     * @param utilityCachePool Utility cache pool.
-     * @param execSvc Executor service.
-     * @param svcExecSvc Services' executor service.
-     * @param sysExecSvc System executor service.
-     * @param stripedExecSvc Striped executor.
-     * @param p2pExecSvc P2P executor service.
-     * @param mgmtExecSvc Management executor service.
-     * @param dataStreamExecSvc data stream executor service.
-     * @param restExecSvc Reset executor service.
-     * @param affExecSvc Affinity executor service.
-     * @param idxExecSvc Indexing executor service.
-     * @param callbackExecSvc Callback executor service.
-     * @param qryExecSvc Query executor service.
-     * @param schemaExecSvc Schema executor service.
-     * @param rebalanceExecSvc Rebalance executor service.
-     * @param rebalanceStripedExecSvc Rebalance striped executor service.
-     * @param customExecSvcs Custom named executors.
-     * @param workersRegistry Worker registry.
      * @throws IgniteCheckedException if fails to register any of the MBeans.
      */
-    public void registerMBeansAfterNodeStarted(
-        ExecutorService utilityCachePool,
-        final ExecutorService execSvc,
-        final ExecutorService svcExecSvc,
-        final ExecutorService sysExecSvc,
-        final StripedExecutor stripedExecSvc,
-        ExecutorService p2pExecSvc,
-        ExecutorService mgmtExecSvc,
-        StripedExecutor dataStreamExecSvc,
-        ExecutorService restExecSvc,
-        ExecutorService affExecSvc,
-        @Nullable ExecutorService idxExecSvc,
-        IgniteStripedThreadPoolExecutor callbackExecSvc,
-        ExecutorService qryExecSvc,
-        ExecutorService schemaExecSvc,
-        ExecutorService rebalanceExecSvc,
-        IgniteStripedThreadPoolExecutor rebalanceStripedExecSvc,
-        @Nullable final Map<String, ? extends ExecutorService> customExecSvcs,
-        WorkersRegistry workersRegistry
-    ) throws IgniteCheckedException {
+    public void registerMBeansAfterNodeStarted() throws IgniteCheckedException {
         if (U.IGNITE_MBEANS_DISABLED)
             return;
 
@@ -197,46 +150,16 @@ public class IgniteMBeansManager {
         MetricsMxBean metricsMxBean = new MetricsMxBeanImpl(ctx.metric(), log);
         registerMBean("Metrics", metricsMxBean.getClass().getSimpleName(), metricsMxBean, MetricsMxBean.class);
 
-        // Executors
-        registerExecutorMBean("GridUtilityCacheExecutor", utilityCachePool);
-        registerExecutorMBean("GridExecutionExecutor", execSvc);
-        registerExecutorMBean("GridServicesExecutor", svcExecSvc);
-        registerExecutorMBean("GridSystemExecutor", sysExecSvc);
-        registerExecutorMBean("GridClassLoadingExecutor", p2pExecSvc);
-        registerExecutorMBean("GridManagementExecutor", mgmtExecSvc);
-        registerExecutorMBean("GridAffinityExecutor", affExecSvc);
-        registerExecutorMBean("GridCallbackExecutor", callbackExecSvc);
-        registerExecutorMBean("GridQueryExecutor", qryExecSvc);
-        registerExecutorMBean("GridSchemaExecutor", schemaExecSvc);
-        registerExecutorMBean("GridRebalanceExecutor", rebalanceExecSvc);
-        registerExecutorMBean("GridRebalanceStripedExecutor", rebalanceStripedExecSvc);
-
-        registerStripedExecutorMBean("GridDataStreamExecutor", dataStreamExecSvc);
-
-        if (idxExecSvc != null)
-            registerExecutorMBean("GridIndexingExecutor", idxExecSvc);
-
-        if (ctx.config().getConnectorConfiguration() != null)
-            registerExecutorMBean("GridRestExecutor", restExecSvc);
-
-        if (stripedExecSvc != null) {
-            // striped executor uses a custom adapter
-            registerStripedExecutorMBean("StripedExecutor", stripedExecSvc);
-        }
-
-        if (customExecSvcs != null) {
-            for (Map.Entry<String, ? extends ExecutorService> entry : customExecSvcs.entrySet())
-                registerExecutorMBean(entry.getKey(), entry.getValue());
-        }
+        ctx.pools().registerMxBeans(this);
 
         if (U.IGNITE_TEST_FEATURES_ENABLED) {
-            WorkersControlMXBean workerCtrlMXBean = new WorkersControlMXBeanImpl(workersRegistry);
+            WorkersControlMXBean workerCtrlMXBean = new WorkersControlMXBeanImpl(ctx.workersRegistry());
 
             registerMBean("Kernal", workerCtrlMXBean.getClass().getSimpleName(),
                 workerCtrlMXBean, WorkersControlMXBean.class);
         }
 
-        FailureHandlingMxBean blockOpCtrlMXBean = new FailureHandlingMxBeanImpl(workersRegistry,
+        FailureHandlingMxBean blockOpCtrlMXBean = new FailureHandlingMxBeanImpl(ctx.workersRegistry(),
             ctx.cache().context().database());
 
         registerMBean("Kernal", blockOpCtrlMXBean.getClass().getSimpleName(), blockOpCtrlMXBean,
@@ -265,28 +188,6 @@ public class IgniteMBeansManager {
             new WarmUpMXBeanImpl(ctx.cache()),
             WarmUpMXBean.class
         );
-    }
-
-    /**
-     * Registers a {@link ThreadPoolMXBean} for an executor.
-     *
-     * @param name name of the bean to register
-     * @param exec executor to register a bean for
-     * @throws IgniteCheckedException if registration fails.
-     */
-    private void registerExecutorMBean(String name, ExecutorService exec) throws IgniteCheckedException {
-        registerMBean("Thread Pools", name, new ThreadPoolMXBeanAdapter(exec), ThreadPoolMXBean.class);
-    }
-
-    /**
-     * Registers a {@link StripedExecutorMXBean} for an striped executor.
-     *
-     * @param name name of the bean to register
-     * @param exec executor to register a bean for
-     * @throws IgniteCheckedException if registration fails.
-     */
-    private void registerStripedExecutorMBean(String name, StripedExecutor exec) throws IgniteCheckedException {
-        registerMBean("Thread Pools", name, new StripedExecutorMXBeanAdapter(exec), StripedExecutorMXBean.class);
     }
 
     /**
