+++ modules/core/src/main/java/org/apache/ignite/internal/processors/metric/GridMetricManager.java
@@ -28,13 +28,8 @@ import java.lang.management.ThreadMXBean;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.RejectedExecutionHandler;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.ThreadPoolExecutor;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 import org.apache.ignite.IgniteCheckedException;
@@ -51,7 +46,6 @@ import org.apache.ignite.internal.processors.metric.impl.DoubleMetricImpl;
 import org.apache.ignite.internal.processors.metric.impl.HistogramMetricImpl;
 import org.apache.ignite.internal.processors.metric.impl.HitRateMetric;
 import org.apache.ignite.internal.processors.timeout.GridTimeoutProcessor;
-import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.future.GridCompoundFuture;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.T2;
@@ -63,11 +57,9 @@ import org.apache.ignite.spi.metric.Metric;
 import org.apache.ignite.spi.metric.MetricExporterSpi;
 import org.apache.ignite.spi.metric.ReadOnlyMetricManager;
 import org.apache.ignite.spi.metric.ReadOnlyMetricRegistry;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.apache.ignite.internal.IgniteNodeAttributes.ATTR_PHY_RAM;
 import static org.apache.ignite.internal.processors.metric.impl.MetricUtils.fromFullName;
 import static org.apache.ignite.internal.processors.metric.impl.MetricUtils.metricName;
@@ -82,52 +74,6 @@ import static org.apache.ignite.internal.util.IgniteUtils.notifyListeners;
 public class GridMetricManager extends GridManagerAdapter<MetricExporterSpi> implements ReadOnlyMetricManager {
     /** Class name for a SQL view metrics exporter. */
     public static final String SQL_SPI = "org.apache.ignite.internal.processors.metric.sql.SqlViewMetricExporterSpi";
-
-    /** */
-    public static final String ACTIVE_COUNT_DESC = "Approximate number of threads that are actively executing tasks.";
-
-    /** */
-    public static final String COMPLETED_TASK_DESC = "Approximate total number of tasks that have completed execution.";
-
-    /** */
-    public static final String CORE_SIZE_DESC = "The core number of threads.";
-
-    /** */
-    public static final String LARGEST_SIZE_DESC = "Largest number of threads that have ever simultaneously been in the pool.";
-
-    /** */
-    public static final String MAX_SIZE_DESC = "The maximum allowed number of threads.";
-
-    /** */
-    public static final String POOL_SIZE_DESC = "Current number of threads in the pool.";
-
-    /** */
-    public static final String TASK_COUNT_DESC = "Approximate total number of tasks that have been scheduled for execution.";
-
-    /** */
-    public static final String QUEUE_SIZE_DESC = "Current size of the execution queue.";
-
-    /** */
-    public static final String KEEP_ALIVE_TIME_DESC = "Thread keep-alive time, which is the amount of time which threads in excess of " +
-        "the core pool size may remain idle before being terminated.";
-
-    /** */
-    public static final String IS_SHUTDOWN_DESC = "True if this executor has been shut down.";
-
-    /** */
-    public static final String IS_TERMINATED_DESC = "True if all tasks have completed following shut down.";
-
-    /** */
-    public static final String IS_TERMINATING_DESC = "True if terminating but not yet terminated.";
-
-    /** */
-    public static final String REJ_HND_DESC = "Class name of current rejection handler.";
-
-    /** */
-    public static final String THRD_FACTORY_DESC = "Class name of thread factory used to create new threads.";
-
-    /** Group for a thread pools. */
-    public static final String THREAD_POOLS = "threadPools";
 
     /** Metrics update frequency. */
     private static final long METRICS_UPDATE_FREQ = 3000;
@@ -574,185 +520,6 @@ public class GridMetricManager extends GridManagerAdapter<MetricExporterSpi> imp
         }
 
         return (T) m;
-    }
-
-    /**
-     * Registers all metrics for thread pools.
-     *
-     * @param utilityCachePool Utility cache pool.
-     * @param execSvc Executor service.
-     * @param svcExecSvc Services' executor service.
-     * @param sysExecSvc System executor service.
-     * @param stripedExecSvc Striped executor.
-     * @param p2pExecSvc P2P executor service.
-     * @param mgmtExecSvc Management executor service.
-     * @param dataStreamExecSvc Data stream executor service.
-     * @param restExecSvc Reset executor service.
-     * @param affExecSvc Affinity executor service.
-     * @param idxExecSvc Indexing executor service.
-     * @param callbackExecSvc Callback executor service.
-     * @param qryExecSvc Query executor service.
-     * @param schemaExecSvc Schema executor service.
-     * @param rebalanceExecSvc Rebalance executor service.
-     * @param rebalanceStripedExecSvc Rebalance striped executor service.
-     * @param customExecSvcs Custom named executors.
-     */
-    public void registerThreadPools(
-        ExecutorService utilityCachePool,
-        ExecutorService execSvc,
-        ExecutorService svcExecSvc,
-        ExecutorService sysExecSvc,
-        StripedExecutor stripedExecSvc,
-        ExecutorService p2pExecSvc,
-        ExecutorService mgmtExecSvc,
-        StripedExecutor dataStreamExecSvc,
-        ExecutorService restExecSvc,
-        ExecutorService affExecSvc,
-        @Nullable ExecutorService idxExecSvc,
-        IgniteStripedThreadPoolExecutor callbackExecSvc,
-        ExecutorService qryExecSvc,
-        ExecutorService schemaExecSvc,
-        ExecutorService rebalanceExecSvc,
-        IgniteStripedThreadPoolExecutor rebalanceStripedExecSvc,
-        @Nullable final Map<String, ? extends ExecutorService> customExecSvcs
-    ) {
-        // Executors
-        monitorExecutor("GridUtilityCacheExecutor", utilityCachePool);
-        monitorExecutor("GridExecutionExecutor", execSvc);
-        monitorExecutor("GridServicesExecutor", svcExecSvc);
-        monitorExecutor("GridSystemExecutor", sysExecSvc);
-        monitorExecutor("GridClassLoadingExecutor", p2pExecSvc);
-        monitorExecutor("GridManagementExecutor", mgmtExecSvc);
-        monitorExecutor("GridAffinityExecutor", affExecSvc);
-        monitorExecutor("GridCallbackExecutor", callbackExecSvc);
-        monitorExecutor("GridQueryExecutor", qryExecSvc);
-        monitorExecutor("GridSchemaExecutor", schemaExecSvc);
-        monitorExecutor("GridRebalanceExecutor", rebalanceExecSvc);
-        monitorExecutor("GridRebalanceStripedExecutor", rebalanceStripedExecSvc);
-
-        monitorStripedPool("GridDataStreamExecutor", dataStreamExecSvc);
-
-        if (idxExecSvc != null)
-            monitorExecutor("GridIndexingExecutor", idxExecSvc);
-
-        if (ctx.config().getConnectorConfiguration() != null)
-            monitorExecutor("GridRestExecutor", restExecSvc);
-
-        if (stripedExecSvc != null) {
-            // Striped executor uses a custom adapter.
-            monitorStripedPool("StripedExecutor", stripedExecSvc);
-        }
-
-        if (customExecSvcs != null) {
-            for (Map.Entry<String, ? extends ExecutorService> entry : customExecSvcs.entrySet())
-                monitorExecutor(entry.getKey(), entry.getValue());
-        }
-    }
-
-    /**
-     * Creates a MetricSet for an executor.
-     *
-     * @param name Name of the bean to register.
-     * @param execSvc Executor to register a bean for.
-     */
-    private void monitorExecutor(String name, ExecutorService execSvc) {
-        MetricRegistry mreg = registry(metricName(THREAD_POOLS, name));
-
-        if (execSvc instanceof ThreadPoolExecutor) {
-            ThreadPoolExecutor exec = (ThreadPoolExecutor)execSvc;
-
-            mreg.register("ActiveCount", exec::getActiveCount, ACTIVE_COUNT_DESC);
-            mreg.register("CompletedTaskCount", exec::getCompletedTaskCount, COMPLETED_TASK_DESC);
-            mreg.register("CorePoolSize", exec::getCorePoolSize, CORE_SIZE_DESC);
-            mreg.register("LargestPoolSize", exec::getLargestPoolSize, LARGEST_SIZE_DESC);
-            mreg.register("MaximumPoolSize", exec::getMaximumPoolSize, MAX_SIZE_DESC);
-            mreg.register("PoolSize", exec::getPoolSize, POOL_SIZE_DESC);
-            mreg.register("TaskCount", exec::getTaskCount, TASK_COUNT_DESC);
-            mreg.register("QueueSize", () -> exec.getQueue().size(), QUEUE_SIZE_DESC);
-            mreg.register("KeepAliveTime", () -> exec.getKeepAliveTime(MILLISECONDS), KEEP_ALIVE_TIME_DESC);
-            mreg.register("Shutdown", exec::isShutdown, IS_SHUTDOWN_DESC);
-            mreg.register("Terminated", exec::isTerminated, IS_TERMINATED_DESC);
-            mreg.register("Terminating", exec::isTerminating, IS_TERMINATING_DESC);
-            mreg.register("RejectedExecutionHandlerClass", () -> {
-                RejectedExecutionHandler hnd = exec.getRejectedExecutionHandler();
-
-                return hnd == null ? "" : hnd.getClass().getName();
-            }, String.class, REJ_HND_DESC);
-            mreg.register("ThreadFactoryClass", () -> {
-                ThreadFactory factory = exec.getThreadFactory();
-
-                return factory == null ? "" : factory.getClass().getName();
-            }, String.class, THRD_FACTORY_DESC);
-        }
-        else {
-            mreg.longMetric("ActiveCount", ACTIVE_COUNT_DESC).value(0);
-            mreg.longMetric("CompletedTaskCount", COMPLETED_TASK_DESC).value(0);
-            mreg.longMetric("CorePoolSize", CORE_SIZE_DESC).value(0);
-            mreg.longMetric("LargestPoolSize", LARGEST_SIZE_DESC).value(0);
-            mreg.longMetric("MaximumPoolSize", MAX_SIZE_DESC).value(0);
-            mreg.longMetric("PoolSize", POOL_SIZE_DESC).value(0);
-            mreg.longMetric("TaskCount", TASK_COUNT_DESC);
-            mreg.longMetric("QueueSize", QUEUE_SIZE_DESC).value(0);
-            mreg.longMetric("KeepAliveTime", KEEP_ALIVE_TIME_DESC).value(0);
-            mreg.register("Shutdown", execSvc::isShutdown, IS_SHUTDOWN_DESC);
-            mreg.register("Terminated", execSvc::isTerminated, IS_TERMINATED_DESC);
-            mreg.longMetric("Terminating", IS_TERMINATING_DESC);
-            mreg.objectMetric("RejectedExecutionHandlerClass", String.class, REJ_HND_DESC).value("");
-            mreg.objectMetric("ThreadFactoryClass", String.class, THRD_FACTORY_DESC).value("");
-        }
-    }
-
-    /**
-     * Creates a MetricSet for an stripped executor.
-     *
-     * @param name name of the bean to register
-     * @param svc Executor.
-     */
-    private void monitorStripedPool(String name, StripedExecutor svc) {
-        MetricRegistry mreg = registry(metricName(THREAD_POOLS, name));
-
-        mreg.register("DetectStarvation",
-            svc::detectStarvation,
-            "True if possible starvation in striped pool is detected.");
-
-        mreg.register("StripesCount",
-            svc::stripesCount,
-            "Stripes count.");
-
-        mreg.register("Shutdown",
-            svc::isShutdown,
-            "True if this executor has been shut down.");
-
-        mreg.register("Terminated",
-            svc::isTerminated,
-            "True if all tasks have completed following shut down.");
-
-        mreg.register("TotalQueueSize",
-            svc::queueSize,
-            "Total queue size of all stripes.");
-
-        mreg.register("TotalCompletedTasksCount",
-            svc::completedTasks,
-            "Completed tasks count of all stripes.");
-
-        mreg.register("StripesCompletedTasksCounts",
-            svc::stripesCompletedTasks,
-            long[].class,
-            "Number of completed tasks per stripe.");
-
-        mreg.register("ActiveCount",
-            svc::activeStripesCount,
-            "Number of active tasks of all stripes.");
-
-        mreg.register("StripesActiveStatuses",
-            svc::stripesActiveStatuses,
-            boolean[].class,
-            "Number of active tasks per stripe.");
-
-        mreg.register("StripesQueueSizes",
-            svc::stripesQueueSizes,
-            int[].class,
-            "Size of queue per stripe.");
     }
 
     /**
